requirements-core.txt
# DARWIN Core 2.0 - Minimal Dependencies
# Modular architecture with external plugin communication

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
hypercorn==0.16.0  # For future HTTP/3 support
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Clients
httpx[http2]==0.26.0
# Note: http3 removed per technical review - use HTTP/2, proxy handles HTTP/3

# gRPC (HTTP/2)
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
protobuf==4.25.1

# Apache Pulsar (Event-Driven)
pulsar-client==3.4.0

# Kubernetes Integration
kubernetes==28.1.0

# OpenTelemetry (Observability)
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-exporter-otlp-proto-grpc==1.22.0

# Semantic Memory
chromadb>=0.5.0  # ChromaDB v2 API with tenant/database support
sentence-transformers==2.3.1

# ML/RL (Continuous Learning)
scikit-learn==1.4.0
numpy==1.26.3

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scientific corpus ingestion
PyPDF2==3.0.1
python-docx==1.1.0

